backend/requirements.txt
# PICKSY — Backend Dependencies
# Installer avec : pip install -r requirements.txt

# Framework API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Base de données
supabase>=2.10.0

# LLM (DeepSeek — compatible SDK OpenAI)
openai>=1.52.0

# Scraping
firecrawl-py>=1.0.0
trafilatura>=2.0.0
ddgs>=0.1.0

# HTML parsing
httpx>=0.27
lxml>=5.0.0
beautifulsoup4>=4.12.0

# Queue & cache
celery[redis]==5.4.0
redis[asyncio]==5.0.0

# Newsletter
resend>=2.0.0

# Templates email
jinja2==3.1.4

# Config & utils
pydantic[email]==2.7.0
pydantic-settings==2.2.0
python-dotenv==1.0.0

# Icecat / parsing
requests==2.31.0

# Correctifs Cicéron
gunicorn==22.0.0
python-json-logger==3.0.0
tenacity==9.0.0
unidecode==1.3.8
